Accept path-like inputs in utils.py
This PR modifies the `read_lines` and `write_lines` functions to accept
[path-like objects](https://www.python.org/dev/peps/pep-0519/), in particular
[pathlib](https://docs.python.org/3/library/pathlib.html)'s paths.

It doesn't really change behavior otherwise - it still just looks for `.json` in
the file extension to decide how to read/write the data. The only difference is
that if the file extension is not `.json` or `.txt`, the functions now treat the
file as plaintext rather than returning `None`.

PEP 519, which established the `.__fspath__()` method being called here, is
Python 3.6+ only, if that's a concern for compatibility. Another file is using
f-strings, which are another Python 3.6 feature, so I don't think this is a big
deal.